move shape classes to new shapes.py
